// src/types/microcms.ts
export type Category = {
  id: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  revisedAt: string;
  name: string;
};

export type Blog = {
  id: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  revisedAt: string;
  title: string;
  content: string;
  category: Category[];
};

export type Blogs = {
  contents: Blog[];
  totalCount: number;
  offset: number;
  limit: number;
};